+++ modules/History.js
@@ -1,3 +1,4 @@
+import warning from 'warning'
 import { history } from './PropTypes'
 
 /**
@@ -10,6 +11,7 @@ const History = {
   },
 
   componentWillMount() {
+    warning(false, 'the `History` mixin is deprecated, please access `context.router` with your own `contextTypes` or see https://github.com/rackt/react-router/blob/v1.1.0/CHANGES.md#v110 for more options')
     this.history = this.context.history
   }